Update file 20150518212122_create_orders.rb

+++ db/migrate/20150518212122_create_orders.rb
@@ -4,7 +4,7 @@ class CreateOrders < ActiveRecord::Migration
       t.belongs_to :customer, index: true
       t.belongs_to :product, index: true
       t.integer :quantity
-      t.integer :total_price
+      t.money :total_price
       t.column :status, :string, default: 'Received'
       t.integer :user_session_id
       t.belongs_to :coupon, index: true
